Add files via upload
Please write a program which asks the user to type in a string and a single character. The program then prints all three character slices which begins with the character specified by the user. You may assume the input string is at least three characters long. The program must print out three characters, or else nothing.